+++ ST_engine/src/main/drawing_manager/drawing_manager.hpp
@@ -75,8 +75,8 @@ class drawing_manager{
         uint8_t darkness_level = 0;
         uint8_t lights_quality = 0;
 
-        size_t default_font_normal = 0;
-        size_t default_font_small = 0;
+        uint16_t default_font_normal = 0;
+    uint16_t default_font_small = 0;
 
         //debug
         bool collisions_shown = false;
