+++ ST_engine/src/main/game_manager/level/level.hpp
@@ -41,7 +41,7 @@ namespace ST {
          * Contains all entities, lights and text objects.
          * Contains the background overlay and the camera.
          */
-        ska::bytell_hash_map<size_t, std::vector<ST::key>> actions_Buttons{};
+        ska::bytell_hash_map<uint16_t , std::vector<ST::key>> actions_Buttons{};
         std::vector<ST::entity> entities{};
         std::vector<ST::light> lights{};
         std::vector<ST::text> text_objects{};
